Allow validator to change value
Allows a validator to change the value, rather than just return an error
message

If a validator (for example onbeforesave="someValidator($data)" returns
an object with the property newValue, the existing value will be updated
to the value of newValue.

Got the idea from jQuery xeditable, and adapted it to work here in
Angular.